+++ CHANGELOG.md
@@ -22,12 +22,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed the name of class `SplineBuliderDerivField2D` to fix typo (->`SplineBuilderDerivField2D`).
 - Update DDC to [v0.12.0](https://github.com/CExA-project/ddc/releases/tag/v0.12.0).
 - Changed FindLAPACKE CMake module to the version in DDC.
+- Changed `BslAdvectionPolar` template parameters and constructor to take a builder and evaluator instead of an `Interpolator2D`.
 
 ### Deprecated
 
 ### Removed
 
 - Remove DDC submodule.
+- Remove 2D `Interpolator` classes:
+  - `IInterpolator2D`
+  - `SplineInterpolator2D`
 
 ## [v0.7.0] - 2026-03-18